+++ cafehandler/precode.go
@@ -1,0 +1,45 @@
+package cafehandler
+
+import (
+	"net/http"
+	"strconv"
+	"strings"
+)
+
+var cafeList = map[string][]string{
+	"moscow": {"Мир кофе", "Сладкоежка", "Кофе и завтраки", "Сытый студент"},
+}
+
+func mainHandle(w http.ResponseWriter, req *http.Request) {
+	countStr := req.URL.Query().Get("count")
+	if countStr == "" {
+		w.WriteHeader(http.StatusBadRequest)
+		w.Write([]byte("count missing"))
+		return
+	}
+
+	count, err := strconv.Atoi(countStr)
+	if err != nil {
+		w.WriteHeader(http.StatusBadRequest)
+		w.Write([]byte("wrong count value"))
+		return
+	}
+
+	city := req.URL.Query().Get("city")
+
+	cafe, ok := cafeList[city]
+	if !ok {
+		w.WriteHeader(http.StatusBadRequest)
+		w.Write([]byte("wrong city value"))
+		return
+	}
+
+	if count > len(cafe) {
+		count = len(cafe)
+	}
+
+	answer := strings.Join(cafe[:count], ",")
+
+	w.WriteHeader(http.StatusOK)
+	w.Write([]byte(answer))
+}
